// cocktails.web/src/api/cocktailsApi/models/cocktailRs.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cocktails Api
 * Cocktails Api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CocktailModel } from './cocktailModel';
import {
    CocktailModelFromJSON,
    CocktailModelFromJSONTyped,
    CocktailModelToJSON,
    CocktailModelToJSONTyped,
} from './cocktailModel';

/**
 * 
 * @export
 * @interface CocktailRs
 */
export interface CocktailRs {
    /**
     * 
     * @type {CocktailModel}
     * @memberof CocktailRs
     */
    item: CocktailModel;
}

/**
 * Check if a given object implements the CocktailRs interface.
 */
export function instanceOfCocktailRs(value: object): value is CocktailRs {
    if (!('item' in value) || value['item'] === undefined) return false;
    return true;
}

export function CocktailRsFromJSON(json: any): CocktailRs {
    return CocktailRsFromJSONTyped(json, false);
}

export function CocktailRsFromJSONTyped(json: any, ignoreDiscriminator: boolean): CocktailRs {
    if (json == null) {
        return json;
    }
    return {
        
        'item': CocktailModelFromJSON(json['item']),
    };
}

export function CocktailRsToJSON(json: any): CocktailRs {
    return CocktailRsToJSONTyped(json, false);
}

export function CocktailRsToJSONTyped(value?: CocktailRs | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'item': CocktailModelToJSON(value['item']),
    };
}