// cocktails.web/src/api/cocktailsApi/models/glasswareTypeModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cocktails Api
 * Cocktails Api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum GlasswareTypeModel {
    None = 'none',
    Rocks = 'rocks',
    Highball = 'highball',
    ShotGlass = 'shotGlass',
    Coupe = 'coupe',
    CopperMug = 'copperMug',
    Collins = 'collins',
    CocktailGlass = 'cocktailGlass',
    WineGlass = 'wineGlass',
    Flute = 'flute',
    Lowball = 'lowball',
    Fizz = 'fizz',
    TikiMug = 'tikiMug',
    PintGlass = 'pintGlass',
    JulepTin = 'julepTin',
    DoubleRocks = 'doubleRocks',
    Hurricane = 'hurricane',
    HollowedPineapple = 'hollowedPineapple',
    Snifter = 'snifter',
    ScorpionBowl = 'scorpionBowl'
}


export function instanceOfGlasswareTypeModel(value: any): boolean {
    for (const key in GlasswareTypeModel) {
        if (Object.prototype.hasOwnProperty.call(GlasswareTypeModel, key)) {
            if (GlasswareTypeModel[key as keyof typeof GlasswareTypeModel] === value) {
                return true;
            }
        }
    }
    return false;
}

export function GlasswareTypeModelFromJSON(json: any): GlasswareTypeModel {
    return GlasswareTypeModelFromJSONTyped(json, false);
}

export function GlasswareTypeModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): GlasswareTypeModel {
    return json as GlasswareTypeModel;
}

export function GlasswareTypeModelToJSON(value?: GlasswareTypeModel | null): any {
    return value as any;
}

export function GlasswareTypeModelToJSONTyped(value: any, ignoreDiscriminator: boolean): GlasswareTypeModel {
    return value as GlasswareTypeModel;
}