// cocktails.web/src/api/cocktailsApi/models/ingredientApplicationModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cocktails Api
 * Cocktails Api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum IngredientApplicationModel {
    Base = 'base',
    Additional = 'additional',
    Garnishment = 'garnishment',
    Muddle = 'muddle'
}


export function instanceOfIngredientApplicationModel(value: any): boolean {
    for (const key in IngredientApplicationModel) {
        if (Object.prototype.hasOwnProperty.call(IngredientApplicationModel, key)) {
            if (IngredientApplicationModel[key as keyof typeof IngredientApplicationModel] === value) {
                return true;
            }
        }
    }
    return false;
}

export function IngredientApplicationModelFromJSON(json: any): IngredientApplicationModel {
    return IngredientApplicationModelFromJSONTyped(json, false);
}

export function IngredientApplicationModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): IngredientApplicationModel {
    return json as IngredientApplicationModel;
}

export function IngredientApplicationModelToJSON(value?: IngredientApplicationModel | null): any {
    return value as any;
}

export function IngredientApplicationModelToJSONTyped(value: any, ignoreDiscriminator: boolean): IngredientApplicationModel {
    return value as IngredientApplicationModel;
}