// cocktails.web/src/api/cocktailsApi/models/ingredientRequirementTypeModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cocktails Api
 * Cocktails Api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Whether or not this ingredient is required ('Required' or 'Optional')
 * @export
 * @enum {string}
 */
export enum IngredientRequirementTypeModel {
    None = 'none',
    Optional = 'optional',
    Required = 'required'
}


export function instanceOfIngredientRequirementTypeModel(value: any): boolean {
    for (const key in IngredientRequirementTypeModel) {
        if (Object.prototype.hasOwnProperty.call(IngredientRequirementTypeModel, key)) {
            if (IngredientRequirementTypeModel[key as keyof typeof IngredientRequirementTypeModel] === value) {
                return true;
            }
        }
    }
    return false;
}

export function IngredientRequirementTypeModelFromJSON(json: any): IngredientRequirementTypeModel {
    return IngredientRequirementTypeModelFromJSONTyped(json, false);
}

export function IngredientRequirementTypeModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): IngredientRequirementTypeModel {
    return json as IngredientRequirementTypeModel;
}

export function IngredientRequirementTypeModelToJSON(value?: IngredientRequirementTypeModel | null): any {
    return value as any;
}

export function IngredientRequirementTypeModelToJSONTyped(value: any, ignoreDiscriminator: boolean): IngredientRequirementTypeModel {
    return value as IngredientRequirementTypeModel;
}