// cocktails.web/src/api/cocktailsApi/models/ingredientTypeModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cocktails Api
 * Cocktails Api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum IngredientTypeModel {
    Herb = 'herb',
    Fruit = 'fruit',
    Juice = 'juice',
    Bitters = 'bitters',
    Syrup = 'syrup',
    Protein = 'protein',
    Flowers = 'flowers',
    Sauce = 'sauce',
    Vegetable = 'vegetable',
    Dilution = 'dilution',
    Beer = 'beer',
    Spirit = 'spirit',
    Liqueur = 'liqueur',
    Wine = 'wine',
    Champagne = 'champagne'
}


export function instanceOfIngredientTypeModel(value: any): boolean {
    for (const key in IngredientTypeModel) {
        if (Object.prototype.hasOwnProperty.call(IngredientTypeModel, key)) {
            if (IngredientTypeModel[key as keyof typeof IngredientTypeModel] === value) {
                return true;
            }
        }
    }
    return false;
}

export function IngredientTypeModelFromJSON(json: any): IngredientTypeModel {
    return IngredientTypeModelFromJSONTyped(json, false);
}

export function IngredientTypeModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): IngredientTypeModel {
    return json as IngredientTypeModel;
}

export function IngredientTypeModelToJSON(value?: IngredientTypeModel | null): any {
    return value as any;
}

export function IngredientTypeModelToJSONTyped(value: any, ignoreDiscriminator: boolean): IngredientTypeModel {
    return value as IngredientTypeModel;
}