// cocktails.web/src/api/cocktailsApi/models/preparationTypeModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cocktails Api
 * Cocktails Api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Any preparation that should be made with this ingredient
 * @export
 * @enum {string}
 */
export enum PreparationTypeModel {
    None = 'none',
    Chilled = 'chilled',
    FreshlySqueezed = 'freshlySqueezed',
    PeeledAndJuiced = 'peeledAndJuiced',
    FreshlyGrated = 'freshlyGrated',
    Quartered = 'quartered',
    FreshPressed = 'freshPressed'
}


export function instanceOfPreparationTypeModel(value: any): boolean {
    for (const key in PreparationTypeModel) {
        if (Object.prototype.hasOwnProperty.call(PreparationTypeModel, key)) {
            if (PreparationTypeModel[key as keyof typeof PreparationTypeModel] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PreparationTypeModelFromJSON(json: any): PreparationTypeModel {
    return PreparationTypeModelFromJSONTyped(json, false);
}

export function PreparationTypeModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): PreparationTypeModel {
    return json as PreparationTypeModel;
}

export function PreparationTypeModelToJSON(value?: PreparationTypeModel | null): any {
    return value as any;
}

export function PreparationTypeModelToJSONTyped(value: any, ignoreDiscriminator: boolean): PreparationTypeModel {
    return value as PreparationTypeModel;
}