// cocktails.web/src/api/cocktailsApi/models/uofMTypeModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cocktails Api
 * Cocktails Api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * The unit of measure when using this ingredient in a cocktail recipe
 * @export
 * @enum {string}
 */
export enum UofMTypeModel {
    None = 'none',
    Ounces = 'ounces',
    Dashes = 'dashes',
    Tablespoon = 'tablespoon',
    Topoff = 'topoff',
    Item = 'item',
    Teaspoon = 'teaspoon',
    ToTaste = 'toTaste',
    Barspoon = 'barspoon',
    Cups = 'cups',
    Splash = 'splash',
    Discretion = 'discretion'
}


export function instanceOfUofMTypeModel(value: any): boolean {
    for (const key in UofMTypeModel) {
        if (Object.prototype.hasOwnProperty.call(UofMTypeModel, key)) {
            if (UofMTypeModel[key as keyof typeof UofMTypeModel] === value) {
                return true;
            }
        }
    }
    return false;
}

export function UofMTypeModelFromJSON(json: any): UofMTypeModel {
    return UofMTypeModelFromJSONTyped(json, false);
}

export function UofMTypeModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): UofMTypeModel {
    return json as UofMTypeModel;
}

export function UofMTypeModelToJSON(value?: UofMTypeModel | null): any {
    return value as any;
}

export function UofMTypeModelToJSONTyped(value: any, ignoreDiscriminator: boolean): UofMTypeModel {
    return value as UofMTypeModel;
}